simulate_sex_rate will now display Ne
This will use the Wapels 89 estimator
of Ne which will give the estimate and
the SE around it.

I've also found out that the statistics
can be evaluated by passing a tuple,
meaning that things that come in lists
(like the heterozygosity at each locus)
does not need to reference each element.