add moved pop2feather from zk_utils
I got the conversion working. The feather format does not save on
space per-se, but it does save on the load time and I have been
able to compress it to ~85% so I can store hundreds of thousands
of lines in a single data set, meaning that I can upload the
data to the CGRB in larger chunks to avoid long read/write times.