update analyze_and_save
I originally had it save the data within the same Rda file, but
I realized that it was MUCH better to save these in separate files
with similar names. This way, I can retrieve the data if I need
to inspect it, but the results are there no matter what and I don't
overload the computer's memory trying to read in everything at once.

-Z: