update Makefile
Note: makefile will complain if a dependency doesn't
exist, but it will not complain for the correct one!

CIP: I had a missing directory for the html rule,
but I only got

"no rule to make target reports/genomic_data_processing.html"

I fixed it by figuring out which directory was
missing and adding it back.